# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "OpenVPN Access Server Management API with Bulk Operations, Advanced Search, and Comprehensive Redis Caching",
        "title": "GoVPN API Enhanced with Advanced Redis Caching",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/openvpn/search/analytics": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get analytics about search usage and patterns",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Get search analytics",
                "parameters": [
                    {
                        "enum": [
                            "day",
                            "week",
                            "month"
                        ],
                        "type": "string",
                        "default": "week",
                        "description": "Analytics period",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search analytics data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/export": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export advanced search results to file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Export search results",
                "parameters": [
                    {
                        "enum": [
                            "csv",
                            "xlsx",
                            "json"
                        ],
                        "type": "string",
                        "default": "csv",
                        "description": "Export format",
                        "name": "format",
                        "in": "query"
                    },
                    {
                        "description": "Search criteria for export",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedUserSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exported search results",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/groups": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search groups with complex filters and sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Advanced group search",
                "parameters": [
                    {
                        "description": "Advanced search criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedGroupSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedGroupSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/quick": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Perform a quick text search across users and groups",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Quick search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "users",
                            "groups",
                            "all"
                        ],
                        "type": "string",
                        "default": "all",
                        "description": "Search type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Result limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Quick search results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/saved": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of saved searches for current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Get saved searches",
                "parameters": [
                    {
                        "enum": [
                            "users",
                            "groups"
                        ],
                        "type": "string",
                        "description": "Filter by search type",
                        "name": "searchType",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include public saved searches",
                        "name": "includePublic",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.SavedSearchResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save frequently used search criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Save search query",
                "parameters": [
                    {
                        "description": "Search to save",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SavedSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.SavedSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/saved/{searchId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a saved search",
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Delete saved search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Saved search ID",
                        "name": "searchId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/saved/{searchId}/execute": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Execute a previously saved search",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Execute saved search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Saved search ID",
                        "name": "searchId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group search results",
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedGroupSearchResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/suggestions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get autocomplete suggestions for search fields",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Get search suggestions",
                "parameters": [
                    {
                        "description": "Search suggestions criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SearchSuggestionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SearchSuggestionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/openvpn/search/users": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search users with complex filters and sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Advanced Search"
                ],
                "summary": "Advanced user search",
                "parameters": [
                    {
                        "description": "Advanced search criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedUserSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdvancedUserSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AdvancedGroupSearchResponse": {
            "type": "object",
            "properties": {
                "filters": {
                    "$ref": "#/definitions/dto.AdvancedGroupSearchRequest"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EnhancedGroupResponse"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 25
                },
                "metadata": {
                    "$ref": "#/definitions/dto.GroupSearchMetadata"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 25
                },
                "totalPages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and RSA256 JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}